requirements.txt
flask
flask-cors
python-dotenv
gunicorn==23.0.0
redis
user-agents
